Day 067: Mkaing Updates Conditional

+++ code/handmade_intrinsics.h
@@ -19,7 +19,7 @@ SquareRoot(real32 Real32) {
 
 inline real32
 AbsoluteValue(real32 Real32) {
-    int32 Result = fabs(Real32);
+    real32 Result = fabsf(Real32);
     return Result;
 }
